requirement.txt
asgiref==3.8.1
Django==5.0.6
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
